+++ stroom-data/stroom-data-store-impl-fs-db/src/test/java/stroom/data/store/impl/fs/db/TestFileSystemVolumeServiceImpl.java
@@ -30,6 +30,7 @@ import stroom.security.mock.MockSecurityContext;
 import stroom.test.common.util.db.DbTestUtil;
 import stroom.test.common.util.test.StroomUnitTest;
 import stroom.util.io.FileUtil;
+import stroom.util.io.PathCreator;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -91,6 +92,7 @@ class TestFileSystemVolumeServiceImpl extends StroomUnitTest {
 
         final FsVolumeDao fsVolumeDao = new FsVolumeDaoImpl(fsDataStoreDbConnProvider);
         final FsVolumeStateDao fsVolumeStateDao = new FsVolumeStateDaoImpl(fsDataStoreDbConnProvider);
+        final PathCreator pathCreator = new PathCreator(() -> tempDir, () -> tempDir);
         volumeService = new FsVolumeService(fsVolumeDao,
                 fsVolumeStateDao,
                 securityContext,
@@ -98,7 +100,7 @@ class TestFileSystemVolumeServiceImpl extends StroomUnitTest {
                 null,
                 null,
                 null,
-                () -> tempDir);
+                pathCreator);
 
 //        volumeService.volumeList = volumeList;
     }