+++ stroom-data/stroom-data-store-util/src/test/java/stroom/data/store/util/TestStreamGrepTool.java
@@ -26,6 +26,7 @@ import stroom.meta.api.MetaProperties;
 import stroom.test.common.util.db.DbTestModule;
 import stroom.test.common.util.db.DbTestUtil;
 import stroom.test.common.util.test.FileSystemTestUtil;
+import stroom.util.io.HomeDirProviderImpl;
 import stroom.util.io.TempDirProviderImpl;
 
 import com.google.inject.Guice;
@@ -53,6 +54,8 @@ class TestStreamGrepTool {
     @Inject
     private FsVolumeService fsVolumeService;
     @Inject
+    private HomeDirProviderImpl homeDirProvider;
+    @Inject
     private TempDirProviderImpl tempDirProvider;
 
     @Mock
@@ -69,6 +72,7 @@ class TestStreamGrepTool {
         injector.injectMembers(this);
 
         // Clear any lingering volumes or data.
+        homeDirProvider.setHomeDir(tempDir);
         tempDirProvider.setTempDir(tempDir);
         final String path = tempDir
                 .resolve("volumes/defaultStreamVolume")