+++ stroom-node/stroom-node-impl/src/main/java/stroom/node/impl/NodeModule.java
@@ -24,7 +24,7 @@ import stroom.node.api.NodeService;
 import stroom.node.shared.Node;
 import stroom.node.shared.NodeResource;
 import stroom.pipeline.writer.ExtendedPathCreator;
-import stroom.pipeline.writer.PathCreator;
+import stroom.util.io.PathCreator;
 import stroom.util.RunnableWrapper;
 import stroom.util.entityevent.EntityEvent;
 import stroom.util.guice.GuiceUtil;