+++ stroom-util/src/main/java/stroom/util/io/HomeDirProvider.java
@@ -1,0 +1,7 @@
+package stroom.util.io;
+
+import javax.inject.Provider;
+import java.nio.file.Path;
+
+public interface HomeDirProvider extends Provider<Path> {
+}