+++ stroom-util/src/main/java/stroom/util/io/PathConfig.java
@@ -9,7 +9,18 @@ import javax.inject.Singleton;
 
 @Singleton
 public class PathConfig extends AbstractConfig {
+    private String home;// = "~/stroom"; //System.getProperty("user.home") + File.separator + ".stroom";
     private String temp = "/tmp";//System.getProperty("java.io.tmpdir");
+
+    @ReadOnly
+    @JsonPropertyDescription("Home folder to write stuff to. Should only be set per node in application property file")
+    public String getHome() {
+        return home;
+    }
+
+    public void setHome(final String home) {
+        this.home = home;
+    }
 
     @ReadOnly
     @JsonPropertyDescription("Temp folder to write stuff to. Should only be set per node in application property file")
@@ -24,7 +35,8 @@ public class PathConfig extends AbstractConfig {
     @Override
     public String toString() {
         return "PathConfig{" +
-                "temp='" + temp + '\'' +
+                "home='" + home + '\'' +
+                ", temp='" + temp + '\'' +
                 '}';
     }
 }